requirements.txt
alembic==1.11.2
aniso8601==9.0.1
annotated-types==0.5.0
anyio==3.7.1
backoff==2.2.1
bcrypt==4.0.1
blacksheep==1.2.18
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
cleanvision==0.3.3
click==8.1.6
cmake==3.27.2
coloredlogs==14.0
contourpy==1.1.0
croniter==1.4.1
cryptography==41.0.3
cycler==0.11.0
dagit==1.4.5
dagster==1.4.5
dagster-graphql==1.4.5
dagster-webserver==1.4.5
docstring-parser==0.15
essentials==1.1.5
essentials-openapi==1.0.8
exceptiongroup==1.1.2
filelock==3.12.3
fonttools==4.42.0
fsspec==2023.6.0
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
gql==3.4.1
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==2.0.2
grpcio==1.57.0
grpcio-health-checking==1.57.0
guardpost==0.0.9
h11==0.14.0
httplib2==0.22.0
httptools==0.6.0
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lit==16.0.6
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.3
paramiko==3.3.1
pendulum==2.1.2
Pillow==10.0.0
protobuf==4.24.0
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.4.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytzdata==2020.1
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==0.10.1
rodi==1.1.3
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.31.0
sympy==1.12
tabulate==0.9.0
tomli==2.0.1
toposort==1.10
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.166
universal-pathlib==0.0.24
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchdog==3.0.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
zipp==3.16.2
